% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extrapolatr-package.R
\docType{package}
\name{extrapolatr}
\alias{extrapolatr}
\alias{extrapolatr-package}
\title{Sequence extrapolation}
\description{
This package provides tools for convergence acceleration of vector
sequences.  This is perhaps unusual for R, but this is a handy
tool to have at one's disposal when doing, for instance,
expectation maximization.
}

