% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iterate.R
\name{fastIterate}
\alias{fastIterate}
\title{Accelerate a point iteration}
\usage{
fastIterate(start, f, method, n.prior, pre.iter, num.accel, num.cycles, ...)
}
\arguments{
\item{start}{the start value}

\item{f}{the function to iterate}

\item{method}{the acceleration function.  must take an m x n
matrix and return an m-vector.}

\item{n.prior}{the number of prior iterations to run before accelerating}

\item{pre.iter}{how many iterations to do before acceleration (per cycle)}

\item{num.accel}{how many iterations to accelerate}

\item{num.cycles}{how many cycles to do?}

\item{...}{additional arguments for f}
}
\description{
Use the techniques from this package to accelerate a point iteration.
}
\examples{
fastIterate(c(0.2, 0.6), function(x) {c(1 - sin(x[[1]]), cos(x[[2]]))},
            rre, 1, 3, 5, 10)
}

