% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/iterate.R
\name{iterate}
\alias{iterate}
\title{Perform a point iteration}
\usage{
iterate(x, f, n, save = TRUE, ...)
}
\arguments{
\item{x}{the start value of the iteration (not saved)}

\item{f}{the function that performs the iteration}

\item{n}{the number of iterations to do}

\item{save}{do we save the iteration history?}

\item{...}{extra arguments for f}
}
\value{
the output of the iteration.  If save is TRUE, it has an
attribute 'history', which gives the iteration history as a
length(x) x n matrix.
}
\description{
Run a point iteration, saving intermediate values.  This is
a convenient building block for convergence acceleration.
}
\examples{
iterate(0.2, cos, 100, save=FALSE)
iterate(c(-1, 1), function(x) { - x }, 100)
}

