% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rre}
\alias{rre}
\title{Perform reduced-rank extrapolation}
\usage{
rre(sequence)
}
\arguments{
\item{sequence}{a matrix, the columns of which are the sequence to
accelerate}
}
\value{
the RRE limit
}
\description{
This method removes any geometric terms from a sequence, leaving
you with the limit.  Note, as in the examples, that it will happily
remove divergent sequences.  This is a feature rather than a bug.
}
\examples{
ind <- seq_len(5)
vals <- 0.5^(ind - 1)
rre(matrix(vals, nrow=1))
rre(rbind(vals, 2 + vals))
}

